Separated wave class types to files
