+++ Assets/Code/Scripts/SceneManageMent/Waves/AudioWave.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Waves;
+
+[CreateAssetMenu(menuName = "Wave/Audio Wave", fileName = "New Audio Wave")]
+public class AudioWave : HostileWave
+{
+    //The transmission audio clip for this wave
+    [SerializeField]
+    public AudioClip RadioClip;
+
+    public override WaveType GetWaveType()
+    {
+        return WaveType.AudioWave;
+    }
+}
