+++ Assets/Code/Scripts/SceneManageMent/Waves/AudioWave.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 327e0d072a3926a46be00b473da0fa7a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
