+++ Assets/Code/Scripts/SceneManageMent/Waves/Wave.cs
@@ -48,59 +48,4 @@ namespace Waves
         // Activates wave properties and returns a list of wave enemies
         internal abstract List<WaveEnemyInfo> GetWaveAction();
     }
-
-    [CreateAssetMenu(menuName = "Wave/Hostile Wave", fileName = "New Hostile Wave")]
-    public class HostileWave : Wave
-    {
-        //A collection of enemy types, each loop, a squad of each type is spawned
-        [SerializeField]
-        public List<WaveEnemyInfo> waveEnemies;
-
-        //The audio loop for this wave
-        [SerializeField]
-        public AudioClip TrackVariation;
-
-        public override WaveType GetWaveType()
-        {
-            return WaveType.HostileWave;
-        }
-
-        internal override List<WaveEnemyInfo> GetWaveAction()
-        {
-            return waveEnemies;
-        }
-
-        public AudioClip GetTrackVariation()
-        {
-            return TrackVariation;
-        }
-    }
-
-    [CreateAssetMenu(menuName = "Wave/Audio Wave", fileName = "New Audio Wave")]
-    public class AudioWave : HostileWave
-    {
-        //The transmission audio clip for this wave
-        [SerializeField]
-        public AudioClip RadioClip;
-
-        public override WaveType GetWaveType()
-        {
-            return WaveType.AudioWave;
-        }
-    }
-
-    [CreateAssetMenu(menuName = "Wave/LevelComplete Wave", fileName = "New LevelComplete Wave")]
-    public class LevelComplete : Wave
-    {
-        public override WaveType GetWaveType()
-        {
-            return WaveType.LevelComplete;
-        }
-
-        internal override List<WaveEnemyInfo> GetWaveAction()
-        {
-            GameStateController.HandleTrigger(StateTrigger.LevelComplete);
-            return null;
-        }
-    }
 }
