+++ Assets/EditorObject/WaveSequences/DrivingSynth/WastelandSoulWave1.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3e1d21df84211914987ec59742c2f0ff, type: 3}
+  m_Name: WastelandSoulWave1
+  m_EditorClassIdentifier: 
+  DLThreshold: 0
+  waveEnemies: []
+  TrackVariation: {fileID: 0}
