+++ Assets/EditorObject/WaveSequences/DrivingSynth/WastelandSoulWave1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b934e6934dafee941b186b6053aad7a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
